backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
python-multipart>=0.0.6
pandas>=2.0.0
numpy>=1.24.3
scipy>=1.10.1
boto3>=1.26.0
websockets>=11.0.0
python-dotenv>=1.0.0
watchdog>=3.0.0
celery>=5.2.7
matplotlib>=3.7.1
scikit-learn>=1.2.2
pytest>=7.3.1
httpx>=0.24.0
